examples/fep_control.GenerativeAgent: sample predicted chances of goal success
Use our predicted observation with our goal to predict the chances (logits) of
successfully accomplishing optimality.  This makes those chances a random
variable whose value can be computed directly from the observation in the
recognition model, which teaches the generative model that its predictions of
success should be accurate, not optimistic.

Signed-off-by: Eli Sennesh <sennesh.e@husky.neu.edu>